+++ lm-api/lm_api/api/schemas/cluster_status.py
@@ -2,6 +2,7 @@
 
 from pendulum.datetime import DateTime
 from lm_api.api.schemas.base import BaseCreateSchema, BaseUpdateSchema
+from pydantic import ConfigDict, field_serializer
 
 
 class ClusterStatusSchema(BaseCreateSchema, BaseUpdateSchema):
@@ -12,6 +13,9 @@ class ClusterStatusSchema(BaseCreateSchema, BaseUpdateSchema):
     cluster_client_id: str
     interval: int
     last_reported: DateTime
+    model_config = ConfigDict(from_attributes=True, arbitrary_types_allowed=True)
 
-    class Config:
-        orm_mode = True
+    @field_serializer("last_reported", when_used="json")
+    @classmethod
+    def serialize_last_reported(cls, value: DateTime) -> str:
+        return value.isoformat().replace("Z", "+00:00")